refactor: add mvc router createMovie, getId,putMovies and scrapper
